Move companion message decoding out of monitor
